requirements.txt
datasets>=2.0.0
torch>=2.0.0
transformers>=4.30.0
pyyaml>=6.0
numpy>=1.20.0

